Make college map size consistent across index, create, and edit forms; store event location as percentage floats

+++ database/migrations/2025_05_22_190800_create_events_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('thumbnail')->nullable();
             $table->string('title');
             $table->text('description');
-            $table->integer('location_x');
-            $table->integer('location_y');
+            $table->float('location_x');
+            $table->float('location_y');
             $table->date('date');
             $table->enum('status', ['upcoming', 'happened', 'cancelled']);
             $table->timestamps();
